Add time prop to Message component
For the time prop, there also have to be a getTime function to format
the time displaying the time in AM/PM formate. The time is inserted
after the message so the message is moved to its own div called
message-content. The scss file is updated to format the message
specifying a minimum and a maximum width while setting the width to its
max content, and also setting the wrap to break the words. There are
also classes for the time and the message-content to format them
accordingly