Add useNextQuery hook
This hook fetches the next set of query results and updates the results
state variable with the new data. It also sets the hasNextPage variable
to indicate whether there are more results available and sets the
nextUserId ref to the last user's ID, if there is such an ID, which can
be used for the next query call.

Finally, it returns all of this data to the component that
calls the hook