Update SearchResults component
First define lastUserId state variable that will be passed to the
useNextQuery hook and it set by the useGetLastRef hook

useNextQuery hook returns nameResults array that is used to set the list
of found users on the page. It also returns hasNextPage and nextUserId
variables that are used by the useGetLastRef hook

useGetLastRef returns a ref that is attached to last DisplayUser compo-
nent. This ref is used to determine if the component is intersecting

handleSearchUser is a cached function. It navigates to the selected user
It is cached because it is passed as a prop to the DisplayUser component
and caching the function avoids the unnecessary re-renders because the
DisplayUser component is memoized.