Import getFollowingStatus, followUser, unfollowUser
These functions have been extracted to the apiCalls. Remove them and
import them from the apiCalls

Also retrieve socket from the context to pass it to the followUser
function. followUser function uses socket to emit an event "sendFollow"
to the socket server

Also import useAxiosConfig to have an axios instance with an interceptor