Change messageNotifications to notifications
Add an effect to remove notification highlights if they are empty

Create a notificationIds object to store names for icons and their
corresponding drop downs. This object is used in the
showSelectedNotifications function to switch between the notification
drop downs

handleMarkAsRead function sends an api request to delete the specific
notification in the db. Also, it removes the notification from the
context

search-username css class will make cursor a pointer

Rename the messageNotification.length into appropriate notification
objects to display in the drop downs