Move post content into the post css class
The outer class of the component is now post-wrapper

This is done so the comment section could be placed under the post
section and each section can set its own padding

Set post-footer to position relative. Now the liker-list can be
accurately positioned above the like counter

Move LikerList into the post-footer so it can be positioned

Uncomment the ChatBubble component. Now when clicked, it toggles the
TextInput component in and out of view

Add like-count className so a margin could be applied to it